cleaner version of scores
